+++ packages/http/src/decorators/metadata.ts
@@ -1,4 +1,0 @@
-export const CONTROLLER_PREFIX = 'pearl:controller:prefix'
-export const CONTROLLER_MIDDLEWARE = 'pearl:controller:middleware'
-export const ROUTE_DEFINITIONS = 'pearl:route:definitions'
-export const INJECTABLE = 'pearl:injectable'